Describe the shell being a different interface to files.
Mention that files & directories created in the shell are no different than files & directories created "normally" using a GUI program.

This can sometimes be a stumbling block for people are new to using a shell.  Their mental model can be that things done inside the shell only exist inside the shell and anything done inside is "trapped" there.

Having someone see that their newly created files exist the same as any other file can be a good way of changing that mental model.